refactor(photo!): rename from `photos` and make `cli` a submodule
This changes the package to be centered around the `photo()` function rather than the cli functionality.
