+++ photo/__testdata__/square.jpg
