+++ photo/__testdata__/wide.jpg
