+++ photo/deno.json
@@ -1,10 +1,10 @@
 {
-  "name": "@tugrulates/photos",
-  "version": "0.1.5",
+  "name": "@tugrulates/photo",
+  "version": "0.0.0",
   "exports": {
-    ".": "./photos.ts",
+    ".": "./photo.ts",
+    "./cli": "./cli.ts",
     "./exif": "./exif.ts",
-    "./photo": "./photo.ts",
     "./testing": "./testing.ts"
   },
   "compile": {
